Move TileElementBase to its own file
